% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{all}
\alias{all}
\title{All individuals}
\format{A data frame with 1,780 rows and 17 variables that is current as of 24 August 2019:
\describe{
 \item{uniqueID}{An indentifier that is unique to each politician, usually just their surname and their year of birth, but occasionally more in cases where that would not be unique.}
 \item{surname}{The surname of the politician.}
 \item{allOtherNames}{The other names of the politician.}
 \item{firstName}{The first name of the politician.}
 \item{commonName}{The common name of the politician, for instance, Ted instead of Edward.}
 \item{displayName}{The combination of common name (when it exists), and first name otherwise, with surname.}
 \item{earlierOrLaterNames}{Just in case they changed their names.}
 \item{title}{Whether the person has a title.}
 \item{gender}{What gender is the person.}
 \item{birthDate}{What is their birthdate?}
 \item{birthYear}{In some cases the specific birthday isn't known, but the birth year is.}
 \item{deathDate}{What date did they die?}
 \item{member}{Where they a member of the house of representatives?}
 \item{senator}{Where they a senator?}
 \item{wikipedia}{The wikipedia link.}
 \item{adb}{The Australian Dictionary of Biography link.}
 \item{comments}{Miscellaneous comments.}
}}
\usage{
all
}
\description{
This is data about every individual who has been elected to the Australian
House of Representatives or the Senate. Each row is an individual, and the
columns are various pieces of information.
}
\keyword{datasets}
