% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{by_division}
\alias{by_division}
\title{Data from Experiwment 1}
\format{A data frame with NNNN rows and NN variables:
\describe{
\item{uniqueID}{HAHAHA}
\item{division}{HAHAHA}
\item{state}{HAHAHA}
\item{electionDate}{HAHAHA}
\item{byElection}{HAHAHA}
\item{reasonCeasedToBeMember}{HAHAHA}
\item{changedSeat}{HAHAHA}
\item{ceasedToBeMemberDate}{HAHAHA}
\item{Comment}{HAHAHA}
}}
\usage{
by_division
}
\description{
This is data from the first experiment ever to try XYZ using Mechanical
Turk workers.
}
\keyword{datasets}
