% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{by_house}
\alias{by_house}
\title{Data from Experigment 1}
\format{A data frame with NNNN rows and NN variables:
\describe{
\item{uniqueID}{HAHAHA}
\item{house}{HAHAHA}
}}
\usage{
by_house
}
\description{
This is data from the first experiment ever to try XYZ using Mechanical
Turk workers.
}
\keyword{datasets}
