% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{by_party}
\alias{by_party}
\title{Data from Experaiment 1}
\format{A data frame with NNNN rows and NN variables:
\describe{
\item{uniqueID}{HAHAHA}
\item{party}{HAHAHA}
\item{partyFrom}{HAHAHA}
\item{partyTo}{HAHAHA}
\item{comment}{HAHAHA}
}}
\usage{
by_party
}
\description{
This is data from the first experiment ever to try XYZ using Mechanical
Turk workers.
}
\keyword{datasets}
