% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{by_state_senators}
\alias{by_state_senators}
\title{Senators by state}
\format{A data frame with 685 rows and 7 variables:
\describe{
\item{uniqueID}{An indentifier that is unique to each politician, usually just their surname and their year of birth, but occasionally more in cases where that would not be unique.}
\item{senatorsState}{The state that they represented.}
\item{senatorFrom}{When they were elected.}
\item{senatorTo}{Their last day of being a senator.}
\item{senatorEndReason}{Why did they stop being a senator?}
\item{sec15Sel}{Was the senator appointed, rather than voted?}
\item{senatorComments}{Miscellaneous comments.}
}}
\usage{
by_state_senators
}
\description{
This provides information about the state that a senator represents, as of 10 September 2019.
}
\keyword{datasets}
