% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cabinet}
\alias{cabinet}
\title{cabinet}
\format{A data frame with 581 rows and 5 variables:
\describe{
 \item{cabinet}{}
 \item{cabinet_number}{}
 \item{cabinet_name}{}
 \item{uniqueID}{}
 \item{cabinet_from}{}
 \item{cabinet_to}{}
 \item{cabinet_comment}{}
}}
\usage{
cabinet
}
\description{
This is a correspondence between uniqueIDs and the ids used by the Australian parliament house. This is a very initial dataset and shouldn't be used.
}
\keyword{datasets}
