% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{list_prime_ministers}
\alias{list_prime_ministers}
\title{List of prime ministers}
\format{A data frame with 30 rows and 2 variables:
\describe{
\item{uniqueID}{An indentifier that is unique to each politician, usually just their surname and their year of birth, but occasionally more in cases where that would not be unique.}
\item{wasPrimeMinister}{A binary indicator as to whether the person was ever prime minister.}
}}
\usage{
list_prime_ministers
}
\description{
This is a list of the Australian prime ministers, as of 10 September 2019.
}
\keyword{datasets}
