% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ministry}
\alias{ministry}
\title{ministry}
\format{A data frame with 581 rows and 5 variables:
\describe{
   {ministry}{}
   {ministry_number}{}
   {ministry_party}{}
   {ministry_title}{}
   {uniqueID}{}
   {ministry_name}{}
   {ministry_from}{}
   {ministry_to}{}
   {ministry_assistant_minister_or_parliamentary_secretary}{}
   {ministry_comment}{}
}}
\usage{
ministry
}
\description{
This is a correspondence between uniqueIDs and the ids used by the Australian parliament house. This is a very initial dataset and shouldn't be used.
}
\keyword{datasets}
