% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{senators_by_state}
\alias{senators_by_state}
\title{Data from Experiwment 1}
\format{A data frame with NNNN rows and NN variables:
\describe{
\item{uniqueID}{HAHAHA}
\item{state}{HAHAHA}
\item{from}{HAHAHA}
\item{to}{HAHAHA}
\item{reasonCeasedToBeSenator}{HAHAHA}
\item{section15Selection}{HAHAHA}
\item{comments}{HAHAHA}
}}
\usage{
senators_by_state
}
\description{
This is data from the first experiment ever to try XYZ using Mechanical
Turk workers.
}
\keyword{datasets}
