% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{uniqueID_to_aphID}
\alias{uniqueID_to_aphID}
\title{uniqueID to aphID correspondence}
\format{A data frame with 1776 rows and 3 variables:
\describe{
\item{uniqueID}{An indentifier that is unique to each politician, usually just their surname and their year of birth, but occasionally more in cases where that would not be unique.}
\item{aphID}{The id used by the Australian parliament house.}
\item{idsComments}{Miscellaneous comments.}
}}
\usage{
uniqueID_to_aphID
}
\description{
This is a correspondence between uniqueIDs and the ids used by the Australian parliament house
}
\keyword{datasets}
