% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{all}
\alias{all}
\title{All individuals}
\format{A data frame with 1,776 rows and 20 variables, as of 31 October 2019:
\describe{
 \item{uniqueID}{uniqueID is usually the surname of the politician and the year that they were born, e.g. Abbott1859. In certain cases this is not enough to uniquely identify them and then we add the first name, e.g. AndersonCharles1897 and AndersonGordon1897. In cases where there is punctuation in the surname, e.g. Ashley-Brown or O'Brien, this has been removed but capitalisation has been retained, so those would become AshleyBrown or OBrien, respectively.}
 \item{surname}{The surname of the politician.}
 \item{allOtherNames}{The other names of the politician.}
 \item{firstName}{The first name of the politician.}
 \item{commonName}{commonName is used to highlight the name that the politician tended to be known as e.g. Ted instead of Edward This is used in displayName which is a politicians surname and their common name (if they had one) or first name e.g. Abbott, Richard. In cases where this would not be unique, e.g. Francis Baker, an additional name has been added.}
 \item{displayName}{The combination of common name (when it exists), and first name otherwise, with surname.}
 \item{earlierOrLaterNames}{earlierOrLaterNames is mostly used to keep track of women changing their names at marriage. Similarly, title is mostly used to keep track of 'Dr', but both have been used inconsistently and should be only used sparingly.}
 \item{title}{Whether the person has a title.}
 \item{gender}{What gender is the person.}
 \item{birthDate}{What is their birthdate?}
 \item{birthYear}{Some politicians don't have a complete birth date, and instead only have a year of birth. In these cases their entry for birthDate will be empty, but they will have a birthYear.}
 \item{birthPlace}{Based on WikiData, this is the town or other place that they were born in.}
 \item{deathDate}{All death dates are complete, but in the case of one politician -- John William Croft -- this has been inputted, as the circumstances and timing (even year) of his death are unknown.}
 \item{member}{member and senator are binary indicator variables used to signify whether the politician was in the lower or upper house. Most politicians are only in one or the other, but some were in both.}
 \item{senator}{One politician in the dataset was neither a senator nor an MP - Heather Elaine Hill. She remains in the dataset because she was elected to the senate, and the need for this dataset to exactly match the AustralianElections one), however her eligibility was challenged and her election was invalidated, so she was never a senator.}
 \item{wasPrimeMinister}{Flag for whether the politician was ever prime minister.}
 \item{wikidataID}{The wikidata ID.}
 \item{wikipedia}{The wikipedia link.}
 \item{adb}{The Australian Dictionary of Biography link.}
 \item{comments}{Miscellaneous comments.}
}}
\usage{
all
}
\description{
This is data about every individual who has been elected to the Australian
House of Representatives or the Senate. Each row is an individual, and the
columns are various pieces of information.
}
\keyword{datasets}
