% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{all_individuals}
\alias{all_individuals}
\title{Data from Experiment 1}
\format{A data frame with NNNN rows and NN variables:
\describe{
 \item{uniqueID}{HAHAHAHAHAHA}
 \item{surname}{HAHAHAHAHAHA}
 \item{allOtherNames}{HAHAHAHAHAHA}
 \item{firstName}{HAHAHAHAHAHA}
 \item{commonName}{HAHAHAHAHAHA}
 \item{displayName}{HAHAHAHAHAHA}
 \item{earlierOrLaterNames}{HAHAHAHAHAHA}
 \item{title}{HAHAHAHAHAHA}
 \item{gender}{HAHAHAHAHAHA}
 \item{birthDate}{HAHAHAHAHAHA}
 \item{birthYear}{HAHAHAHAHAHA}
 \item{deathDate}{HAHAHAHAHAHA}
 \item{member}{HAHAHAHAHAHA}
 \item{senator}{HAHAHAHAHAHA}
 \item{wikipedia}{HAHAHAHAHAHA}
 \item{adb}{HAHAHAHAHAHA}
 \item{comments}{HAHAHAHAHAHA}
}}
\usage{
all_individuals
}
\description{
This is data from the first experiment ever to try XYZ using Mechanical
Turk workers.
}
\keyword{datasets}
