% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{by_division_mps}
\alias{by_division_mps}
\title{Data by division for members.}
\format{A data frame with 1,418 rows and 9 variables:
\describe{
\item{uniqueID}{An indentifier that is unique to each politician, usually just their surname and their year of birth, but occasionally more in cases where that would not be unique.}
\item{division}{The name of the division.}
\item{state}{The state of the division.}
\item{byElection}{Whether the member was elected in a by-election.}
\item{from}{When did the member start representing that division.}
\item{to}{When did the member finish representing that division.}
\item{reasonCeased}{Why did they stop?}
\item{changedSeat}{Was it just that they changed seats?}
\item{comments}{Miscellaneous comments.}
}}
\usage{
by_division_mps
}
\description{
This is data about the division that a particular member represents.
}
\keyword{datasets}
