% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{by_state_senators}
\alias{by_state_senators}
\title{Senators by state}
\format{A data frame with 684 rows and 7 variables:
\describe{
\item{uniqueID}{An indentifier that is unique to each politician, usually just their surname and their year of birth, but occasionally more in cases where that would not be unique.}
\item{state}{The state that they represented.}
\item{from}{When they were elected.}
\item{to}{Their last day of being a senator.}
\item{reasonCeasedToBeSenator}{Why did they stop being a senator?}
\item{section15Selection}{Was the senator appointed, rather than voted?}
\item{comments}{Miscellaneous comments.}
}}
\usage{
by_state_senators
}
\description{
This provides information about the state that a senator represents, as of 24 August 2019.
}
\keyword{datasets}
