% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{list_prime_ministers}
\alias{list_prime_ministers}
\title{Data from Exdperiment 1}
\format{A data frame with NNNN rows and NN variables:
\describe{
\item{uniqueID}{HAHAHA}
\item{wasPrimeMinister}{HAHAHA}
}}
\usage{
list_prime_ministers
}
\description{
This is data from the first experiment ever to try XYZ using Mechanical
Turk workers.
}
\keyword{datasets}
